requirements.txt
asgiref==3.4.1
attrs==22.2.0
backports.zoneinfo==0.2.1
black==22.8.0
click==8.0.4
dj-database-url==2.2.0
Django==3.2.25
django-cors-headers==3.10.1
django-environ==0.11.2
django-rest-framework==0.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==4.4.0
importlib-metadata==4.8.3
importlib-resources==5.4.0
iniconfig==1.1.1
mypy-extensions==1.0.0
packaging==21.3
pathspec==0.9.0
platformdirs==2.4.0
pluggy==1.0.0
psycopg2==2.9.8
psycopg2-binary==2.9.8
py==1.11.0
PyJWT==2.4.0
pyparsing==3.0.7
pytest-django==4.5.2
python-decouple==3.8
pytz==2024.1
sqlparse==0.4.4
tomli==1.2.3
typed-ast==1.5.5
typing-extensions==4.1.1
zipp==3.6.0
